Add config option to set RTSP Server Address
This can be used to make RPOS return a different IP address in the GetStreamUri onvif message when the RTPS server is running on a different machine.
Note that at the time of this commit RPOS will not know the video width or height or frame rate of the stream coming from a remote RTSP server and will use width/height/framerate values from the device RPOS is running on.