Add username and password athentication. Note:- XML reply wrong
The XML reply when the password is wrong is not in the correct
ONVIF format. This needs to be fixed.
But the code does check and validate the password.

Did not use the soap package's code to generate a digest password.
That only worked if the Nonce was all ascii characters.

ToDo- look at the timestamp the password digest was generated and make
sure it is within 20 seconds of the current time (to avoid replay attacks)