Fix username/password checks - Milestone XProtect warning
There was a case where we would allow ONVIF connections without a username/password (ws-security) even when there was a username/password in the config file.
We now check username in the soap request event instead of the soap Header Found request.

Thankyou Milestone for giving out unlimited trials of upto 8 cameras.
XProtect 2020 R3 reported this error when trying to add RPOS to XProtect