Pass SetPreset through to remove ONVIF Camera. Sony hack.
Sony SRG-XP1 ONVIF API allows you to Set a Preset with the same Preset Name
as existing presets. This shoud be fine as there is a unique Preset Token.
Except the IndigoVision 16.0 VMS parses the list of Preset Names and gets confused
if the name is used twice.

So, SetPreset needs to call GetPreset first to make sure the Preset Name is unique.